pyproject.toml
[tool.poetry]
name = "text-summarizer"
version = "0.1.0"
description = ""
authors = ["Arda Can Aras <ardaaras99@gmail.com>"]
readme = "README.md"
packages = [{ include = "text-summarizer" }]

[tool.poetry.dependencies]
python = "^3.11"
transformers = { extras = ["sentencepiece"], version = "^4.31.0" }
datasets = "^2.14.2"
sacrebleu = "^2.3.1"
rouge-score = "^0.1.2"
py7zr = "^0.20.5"
pandas = "^2.0.3"
nltk = "^3.8.1"
tqdm = "^4.65.0"
pyyaml = "^6.0.1"
matplotlib = "^3.7.2"
torch = "^2.0.1"
ipykernel = "^6.25.0"
boto3 = "^1.28.16"
mypy-boto3-s3 = "^1.28.15.post1"
python-box = "6.0.2"
ensure = "1.0.2"
fastapi = "0.78.0"
uvicorn = "0.18.3"
jinja2 = "3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
